slight refactoring - move webhook handling logic from api to service
