+++ src/server/api/webhook.js
@@ -1,180 +1,8 @@
 // module
-const github = require('../services/github')
-const repoService = require('../services/repo')
-const url = require('../services/url')
-const logger = require('../services/logger')
-
+const webhookService = require('../services/webhook')
 class WebhookApi {
-    async _getHook(owner, repo, noCache, token) {
-        if (!owner || !token) {
-            throw 'Owner/org and token is required.'
-        }
-        let args = {
-            fun: 'listHooks',
-            arg: {
-                noCache: noCache
-            },
-            token: token
-        }
-        if (repo) {
-            args.obj = 'repos'
-            args.arg.owner = owner
-            args.arg.repo = repo
-        } else {
-            args.obj = 'orgs'
-            args.arg.org = owner
-        }
-        const hooks = await github.call(args)
-        let hook = null
-
-        if (hooks && hooks.data && hooks.data.length > 0) {
-            hooks.data.forEach(function (webhook) {
-                if (webhook.active && webhook.config.url && webhook.config.url.indexOf(url.baseWebhook) > -1) {
-                    hook = webhook
-                }
-            })
-        }
-        return hook
-    }
-
-    async _getRepoHook(owner, repo, noCache, token) {
-        let hook = await this._getHook(owner, repo, noCache, token)
-        if (!hook) {
-            try {
-                hook = await this._getHook(owner, undefined, noCache, token)
-            } catch (error) {
-                if (error && error.status !== 404) {
-                    // When the owner is not an org, github returns 404.
-                    throw new Error(error)
-                }
-            }
-        }
-        return hook
-    }
-
-    _getOrgHook(org, noCache, token) {
-        return this._getHook(org, undefined, noCache, token)
-    }
-
-    async _createHook(owner, repo, token) {
-        if (!owner || !token) {
-            throw 'Owner/org and token are required.'
-        }
-        let args = {
-            fun: 'createHook',
-            arg: {
-                noCache: true,
-                config: {
-                    content_type: 'json'
-                },
-                name: 'web',
-                events: ['pull_request'],
-                active: true
-            },
-            token: token
-        }
-        if (repo) {
-            args.obj = 'repos'
-            args.arg.repo = repo
-            args.arg.owner = owner
-            args.arg.config.url = url.webhook(repo)
-        } else {
-            args.obj = 'orgs'
-            args.arg.org = owner
-            args.arg.config.url = url.webhook(owner)
-        }
-        try {
-            const res = await github.call(args)
-            return res.data
-        } catch (error) {
-            logger.info(new Error(error).stack)
-        }
-
-    }
-
-    async _createRepoHook(owner, repo, token) {
-        const hook = await this._getRepoHook(owner, repo, true, token)
-        if (hook) {
-            return hook
-        }
-        return this._createHook(owner, repo, token)
-    }
-
-    async _createOrgHook(org, token) {
-        let hook = await this._getOrgHook(org, true, token)
-        if (hook) {
-            throw new Error('Webhook already exist with base url ' + url.baseWebhook)
-        }
-        hook = await this._createHook(org, undefined, token)
-        return this._handleHookForLinkedRepoInOrg(org, token, this._removeRepoHook.bind(this))
-    }
-
-    async _removeHook(owner, repo, hookId, token) {
-        if (!owner || !token) {
-            throw 'Owner/org and token is required.'
-        }
-        let args = {
-            fun: 'deleteHook',
-            arg: {
-                hook_id: hookId,
-                noCache: true
-            },
-            token: token
-        }
-        if (repo) {
-            args.obj = 'repos'
-            args.arg.owner = owner
-            args.arg.repo = repo
-        } else {
-            args.obj = 'orgs'
-            args.arg.org = owner
-        }
-        try {
-            return github.call(args)
-        } catch (error) {
-            logger.info(new Error(error).stack)
-        }
-
-    }
-
-    async _removeRepoHook(owner, repo, token) {
-        const hook = await this._getRepoHook(owner, repo, true, token)
-        if (!hook) {
-            throw 'No webhook found with base url ' + url.baseWebhook
-        }
-        if (hook.type === 'Organization') {
-            return null
-        }
-        return this._removeHook(owner, repo, hook.id, token)
-    }
-
-    async _removeOrgHook(org, token) {
-        const hook = await this._getOrgHook(org, true, token)
-        if (!hook) {
-            throw 'No webhook found with base url ' + url.baseWebhook
-        }
-        await this._removeHook(org, undefined, hook.id, token)
-        await this._handleHookForLinkedRepoInOrg(org, token, this._createRepoHook.bind(this))
-        return hook
-    }
-
-    async _handleHookForLinkedRepoInOrg(org, token, delegateFun) {
-        const repos = await repoService.getByOwner(org)
-        if (!repos || repos.length === 0) {
-            throw 'No repos found for the org'
-        }
-        const promises = repos.map(repo => {
-            if (!repo.gist) {
-                // Repos with Null CLA will NOT have webhook
-                return null
-            }
-            return delegateFun(repo.owner, repo.repo, token)
-        })
-        return Promise.all(promises)
-    }
-
     async get(req) {
-        return req.args && req.args.org ? this._getOrgHook(req.args.org, req.args.noCache, req.user.token) : this._getRepoHook(req.args.owner, req.args.repo, req.args.noCache, req.user.token)
+        return req.args && req.args.org ? webhookService.getOrgHook(req.args.org, req.args.noCache, req.user.token) : webhookService.getRepoHook(req.args.owner, req.args.repo, req.args.noCache, req.user.token)
 
         // now we will have to check two things:
         // 1) webhook user still has push access to this repo
@@ -187,11 +15,11 @@ class WebhookApi {
     }
 
     async create(req) {
-        return req.args && req.args.orgId ? this._createOrgHook(req.args.org, req.user.token) : this._createRepoHook(req.args.owner, req.args.repo, req.user.token)
+        return req.args && req.args.orgId ? webhookService.createOrgHook(req.args.org, req.user.token) : webhookService.createRepoHook(req.args.owner, req.args.repo, req.user.token)
     }
 
     async remove(req) {
-        return req.args && req.args.org ? this._removeOrgHook(req.args.org, req.user.token) : this._removeRepoHook(req.args.owner, req.args.repo, req.user.token)
+        return req.args && req.args.org ? webhookService.removeOrgHook(req.args.org, req.user.token) : webhookService.removeRepoHook(req.args.owner, req.args.repo, req.user.token)
     }
 }
